+++ src/components/index.js
@@ -3,3 +3,4 @@ export {PostCard} from "./postCard/PostCard";
 export {SuggestionSidebar} from "./SuggestionSidebar";
 export {EditProfileModal} from "./EditProfileModal";
 export {PostModal} from "./PostModal";
+export {FilterRow} from "./FilterRow";
