+++ src/pages/Explore.jsx
@@ -1,15 +1,20 @@
-import { Fragment } from "react";
+import { Fragment, useState } from "react";
 import { useSelector } from "react-redux";
 import { VStack, Text } from "@chakra-ui/react";
-import { PostCard } from "components";
+import { PostCard, FilterRow } from "components";
+import { getFilteredPosts } from "helpers/filterHelpers";
 
 export const Explore = () => {
+  const [sortByValue, setSortByValue] = useState("");
   const { posts } = useSelector((state) => state.posts);
+
+  const filteredPosts = getFilteredPosts(posts, sortByValue);
 
   return (
     <VStack gap={5}>
-      {posts.length > 0 ? (
-        posts.map((item) => (
+       <FilterRow sortByValue={sortByValue} setSortByValue={setSortByValue} />
+      {filteredPosts.length > 0 ? (
+        filteredPosts.map((item) => (
           <Fragment key={item._id}>
             <PostCard postData={item} />
           </Fragment>
